+++ src/Entities/Message.php
@@ -25,4 +25,64 @@ abstract class Message implements EntityInterface
         $this->cookies = array_filter($this->cookies, fn($cookie) => $cookie instanceof Cookie);
         $this->headers = array_filter($this->headers, fn($header) => $header instanceof Header);
     }
+
+    /**
+     * Get HTTP version.
+     *
+     * @return string
+     */
+    public function getHttpVersion(): string
+    {
+        return $this->httpVersion;
+    }
+
+    /**
+     * Get cookies.
+     *
+     * @return array
+     */
+    public function getCookies(): array
+    {
+        return $this->cookies;
+    }
+
+    /**
+     * Get headers.
+     *
+     * @return array
+     */
+    public function getHeaders(): array
+    {
+        return $this->headers;
+    }
+
+    /**
+     * Get headers size.
+     *
+     * @return int
+     */
+    public function getHeadersSize(): int
+    {
+        return $this->headersSize;
+    }
+
+    /**
+     * Get body size.
+     *
+     * @return int
+     */
+    public function getBodySize(): int
+    {
+        return $this->bodySize;
+    }
+
+    /**
+     * Get comment.
+     *
+     * @return string|null
+     */
+    public function getComment(): ?string
+    {
+        return $this->comment;
+    }
 }