+++ src/Entities/Request.php
@@ -121,36 +121,6 @@ class Request extends Message
     }
 
     /**
-     * Get HTTP version.
-     *
-     * @return string
-     */
-    public function getHttpVersion(): string
-    {
-        return $this->httpVersion;
-    }
-
-    /**
-     * Get cookies.
-     *
-     * @return array
-     */
-    public function getCookies(): array
-    {
-        return $this->cookies;
-    }
-
-    /**
-     * Get headers.
-     *
-     * @return array
-     */
-    public function getHeaders(): array
-    {
-        return $this->headers;
-    }
-
-    /**
      * Get query string.
      *
      * @return array
@@ -168,35 +138,5 @@ class Request extends Message
     public function getPostData(): ?PostData
     {
         return $this->postData;
-    }
-
-    /**
-     * Get headers size.
-     *
-     * @return int
-     */
-    public function getHeadersSize(): int
-    {
-        return $this->headersSize;
-    }
-
-    /**
-     * Get body size.
-     *
-     * @return int
-     */
-    public function getBodySize(): int
-    {
-        return $this->bodySize;
-    }
-
-    /**
-     * Get comment.
-     *
-     * @return string|null
-     */
-    public function getComment(): ?string
-    {
-        return $this->comment;
     }
 }